fix+chore: added logs and fix some stuff

+++ src/kernel/munix-core/pmm.h
@@ -5,12 +5,12 @@
 
 typedef struct
 {
-    size_t size;
-    size_t last_used;
+    uint64_t size;
+    uint64_t last_used;
     uint8_t *bitmap;
 } PmmBitmap;
 
 void pmm_init(void);
 Alloc pmm_acquire(void);
 void pmm_release(Alloc *self);
-size_t pmm_available_pages(void);
+uint64_t pmm_available_pages(void);
